pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "异步优化版 RAG 业务实现 - 增强资源管理与生命周期控制"
requires-python = ">=3.9"
dependencies = [
    "aiomysql>=0.2.0",
    "beautifulsoup4>=4.12.3",
    "fastapi==0.109.0",
    "httpx>=0.27.0",
    "langchain-community>=0.3.31",
    "langchain-ollama>=0.3.10",
    "langchain-text-splitters>=0.3.11",
    "loguru>=0.7.3",
    "lxml>=6.0.3",
    "nest-asyncio>=1.6.0",
    "numpy>=1.26.4",
    "openai>=1.50.0",
    "pillow>=11.0.0",
    "pydantic>=2.12.5",
    "pymilvus>=2.6.11",
    "PyYAML>=6.0.1",
    "requests>=2.32.5",
    "scikit-learn>=1.6.1",
    "timm>=1.0.26",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
]